Apply clippy --fix fixes
This commit does not contain everything that `cargo clippy --fix`
changed on Windows. But the least controversial and simplest subset.
The remaining fixes I felt could use some manual cleanup also.

+++ talpid-openvpn/src/wintun.rs
@@ -125,7 +125,7 @@ impl WintunAdapter {
         let mut guid = mem::MaybeUninit::zeroed();
         let result = unsafe { ConvertInterfaceLuidToGuid(&self.luid(), guid.as_mut_ptr()) };
         if result != NO_ERROR as i32 {
-            return Err(io::Error::from_raw_os_error(result as i32));
+            return Err(io::Error::from_raw_os_error(result));
         }
         Ok(unsafe { guid.assume_init() })
     }
